# [GR-44098] Add lambda methods test in debuginfotest