--- a/substratevm/src/com.oracle.svm.test/src/hello/Hello.java
+++ b/substratevm/src/com.oracle.svm.test/src/hello/Hello.java
@@ -205,7 +205,11 @@ public class Hello {
         System.out.println(String.format("t = %s\n", t));
     }
 
-    final static java.util.function.Supplier<String> lambda = () -> "lambdaText";
+    private static java.util.function.Supplier<String> lambda = () -> {
+        StringBuilder sb = new StringBuilder("lambda");
+        sb.append("Text");
+        return sb.toString();
+    };
 
     /* Add new methods above main */
     public static void main(String[] args) {
@@ -229,14 +233,17 @@ public class Hello {
         noInlineThis();
         inlineFrom();
         noInlineManyArgs(0, 1, 2, 3, true, 5, 6, 7, 8, 9,
-                     0.0F, 1.125F, 2.25F, 3.375F, 4.5F, 5.625F, 6.75F, 7.875F, 9.0F, 10.125D, false, 12.375F);
+                        0.0F, 1.125F, 2.25F, 3.375F, 4.5F, 5.625F, 6.75F, 7.875F, 9.0F, 10.125D, false, 12.375F);
         noInlinePassConstants();
+        System.out.println(lambda.get());
         // create and manipulate some foreign types
         CStructTests tests = new CStructTests();
         tests.composite();
         tests.weird();
         System.exit(0);
     }
-    /* Keep main the last method in the file and add new methods right above it to avoid updating all line numbers
-       in testhello.py each time new methods are added */
+    /*
+     * Keep main the last method in the file and add new methods right above it to avoid updating
+     * all line numbers in testhello.py each time new methods are added
+     */
 }