Catches cases like https://github.com/graalvm/mandrel/issues/355 caused by [JDK-8281266](https://bugs.openjdk.java.net/browse/JDK-8281266)