@zakkak 2022-03-23
The underlying issue has been [fixed in jdk17u-dev](https://github.com/openjdk/jdk17u-dev/pull/274) and is expected to land in OpenJDK 17.0.4 (July 19 2022). As a result we could consider this PR for inclusion in GraalVM 22.2.0 (July 26, 2022).