@olpaw 2022-09-20
Hi @zakkak,
while already at adding a test for lambdas it would be great to also test setting breakpoints in lambdas via

- a line breakpoint in the body of the lambda function
- method breakpoint on the name of the lambda function